requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-api-python-client==2.188.0
openai==2.15.0
dateparser==1.2.2
python-dateutil==2.9.0